# [neighbors] Fix iter3() error on Python 3.7 and pass flake8