Fixes #1160 

Tested with tests in #1184 which reproduced the bug.

**Problem description**: Relying on `StopIteration` exceptions bubbling out of generator functions is incorrect since Python 3.7, because these are transformed into `RuntimeError` exceptions. [More info on PEP 479](https://www.python.org/dev/peps/pep-0479/#explanation-of-generators-iterators-and-stopiteration).