@Lucas-C 2019-09-07
Awesome !
I'll merge it after #1184
Does the tests in #1184 trigger the bug under Python 3.7