@davidag 2019-09-08
Yes, the tests you made triggered the problem :+1: I also tested the solution with the repository mentioned in #1160 :relaxed: