@jorgesumle 2019-09-08
The tests from my repository pass now with this code and even the generation got a bit faster.

- Old code:
```
real	2m25,598s
user	2m56,643s
sys	0m10,223s
```
- New code:
```
real	2m6,866s
user	2m54,753s
sys	0m9,101s
```